requirements.txt
Django==3.1.6
django-ckeditor==6.0.0
django-crontab==0.7.1
django-hitcount==1.3.3
Pillow==8.1.0